--- a/CHANGES/5461.misc
+++ b/CHANGES/5461.misc
@@ -1,0 +1,1 @@
+Improve performance at content relation time