@daviddavis 2019-11-11
@fabricio-aguiar this change looks good. did you test it out? what sort of improvement did you see in performance?