@fabricio-aguiar 2019-11-11
> @fabricio-aguiar this change looks good. did you test it out? what sort of improvement did you see in performance?

I tested and the bulk creation worked, but I haven't done a benchmark test, how do you think I can test it? Running performance tests with/without this change?