@fabricio-aguiar 2019-11-11
From 975.89s (0:16:15)  to 912.02s (0:15:12)
```
###############################################
This branch:
###############################################

(pulp) [vagrant@pulp3-source-fedora30 pulp_rpm]$ pytest -vv -r sx --color=yes --pyargs --capture=no --durations=0 pulp_rpm.tests.performance.test_sync
==================================================================================== test session starts =====================================================================================
platform linux -- Python 3.7.3, pytest-5.2.2, py-1.8.0, pluggy-0.13.0 -- /usr/local/lib/pulp/bin/python3
cachedir: .pytest_cache
rootdir: /home/vagrant/devel/pulp_rpm
collected 5 items                                                                                                                                                                            

pulp_rpm/tests/performance/test_sync.py::SyncTestCase::test_centos7_on_demand 
->     Sync => Waiting time (s): 0.058297 | Service time (s): 100.714741

->  Re-sync => Waiting time (s): 0.07064 | Service time (s): 86.760875
PASSED
pulp_rpm/tests/performance/test_sync.py::SyncTestCase::test_centos8_appstream_on_demand 
->     Sync => Waiting time (s): 0.069612 | Service time (s): 83.74244

->  Re-sync => Waiting time (s): 0.072455 | Service time (s): 51.783463
PASSED
pulp_rpm/tests/performance/test_sync.py::SyncTestCase::test_centos8_baseos_on_demand 
->     Sync => Waiting time (s): 0.062542 | Service time (s): 128.680739

->  Re-sync => Waiting time (s): 0.117085 | Service time (s): 96.635704
PASSED
pulp_rpm/tests/performance/test_sync.py::SyncTestCase::test_centos8_kickstart_appstream_on_demand 
->     Sync => Waiting time (s): 0.063477 | Service time (s): 54.69628

->  Re-sync => Waiting time (s): 0.077028 | Service time (s): 50.76915
PASSED
pulp_rpm/tests/performance/test_sync.py::SyncTestCase::test_centos8_kickstart_baseos_on_demand 
->     Sync => Waiting time (s): 0.069305 | Service time (s): 74.913487

->  Re-sync => Waiting time (s): 0.073562 | Service time (s): 68.859118
PASSED

=================================================================================== slowest test durations ===================================================================================
244.39s call     pulp_rpm/tests/performance/test_sync.py::SyncTestCase::test_centos8_baseos_on_demand
222.58s call     pulp_rpm/tests/performance/test_sync.py::SyncTestCase::test_centos7_on_demand
164.82s call     pulp_rpm/tests/performance/test_sync.py::SyncTestCase::test_centos8_appstream_on_demand
151.76s call     pulp_rpm/tests/performance/test_sync.py::SyncTestCase::test_centos8_kickstart_baseos_on_demand
127.80s call     pulp_rpm/tests/performance/test_sync.py::SyncTestCase::test_centos8_kickstart_appstream_on_demand
0.44s setup    pulp_rpm/tests/performance/test_sync.py::SyncTestCase::test_centos7_on_demand
0.00s teardown pulp_rpm/tests/performance/test_sync.py::SyncTestCase::test_centos8_kickstart_baseos_on_demand
0.00s setup    pulp_rpm/tests/performance/test_sync.py::SyncTestCase::test_centos8_appstream_on_demand
0.00s setup    pulp_rpm/tests/performance/test_sync.py::SyncTestCase::test_centos8_kickstart_baseos_on_demand
0.00s setup    pulp_rpm/tests/performance/test_sync.py::SyncTestCase::test_centos8_baseos_on_demand
0.00s setup    pulp_rpm/tests/performance/test_sync.py::SyncTestCase::test_centos8_kickstart_appstream_on_demand
0.00s teardown pulp_rpm/tests/performance/test_sync.py::SyncTestCase::test_centos8_kickstart_appstream_on_demand
0.00s teardown pulp_rpm/tests/performance/test_sync.py::SyncTestCase::test_centos7_on_demand
0.00s teardown pulp_rpm/tests/performance/test_sync.py::SyncTestCase::test_centos8_appstream_on_demand
0.00s teardown pulp_rpm/tests/performance/test_sync.py::SyncTestCase::test_centos8_baseos_on_demand
=============================================================================== 5 passed in 912.02s (0:15:12) ================================================================================

###############################################
master:
###############################################

(pulp) [vagrant@pulp3-source-fedora30 pulp_rpm]$ pytest -vv -r sx --color=yes --pyargs --capture=no --durations=0 pulp_rpm.tests.performance.test_sync
==================================================================================== test session starts =====================================================================================
platform linux -- Python 3.7.3, pytest-5.2.2, py-1.8.0, pluggy-0.13.0 -- /usr/local/lib/pulp/bin/python3
cachedir: .pytest_cache
rootdir: /home/vagrant/devel/pulp_rpm
collected 5 items                                                                                                                                                                            

pulp_rpm/tests/performance/test_sync.py::SyncTestCase::test_centos7_on_demand 
->     Sync => Waiting time (s): 0.056457 | Service time (s): 114.693866

->  Re-sync => Waiting time (s): 0.05682 | Service time (s): 104.678411
PASSED
pulp_rpm/tests/performance/test_sync.py::SyncTestCase::test_centos8_appstream_on_demand 
->     Sync => Waiting time (s): 0.064861 | Service time (s): 93.901083

->  Re-sync => Waiting time (s): 0.075517 | Service time (s): 62.701222
PASSED
pulp_rpm/tests/performance/test_sync.py::SyncTestCase::test_centos8_baseos_on_demand 
->     Sync => Waiting time (s): 0.055392 | Service time (s): 119.431556

->  Re-sync => Waiting time (s): 0.071613 | Service time (s): 104.103759
PASSED
pulp_rpm/tests/performance/test_sync.py::SyncTestCase::test_centos8_kickstart_appstream_on_demand 
->     Sync => Waiting time (s): 0.059198 | Service time (s): 50.945047

->  Re-sync => Waiting time (s): 0.078184 | Service time (s): 53.64913
PASSED
pulp_rpm/tests/performance/test_sync.py::SyncTestCase::test_centos8_kickstart_baseos_on_demand 
->     Sync => Waiting time (s): 0.057699 | Service time (s): 80.730246

->  Re-sync => Waiting time (s): 0.069847 | Service time (s): 68.996173
PASSED

=================================================================================== slowest test durations ===================================================================================
256.42s call     pulp_rpm/tests/performance/test_sync.py::SyncTestCase::test_centos7_on_demand
244.11s call     pulp_rpm/tests/performance/test_sync.py::SyncTestCase::test_centos8_baseos_on_demand
188.42s call     pulp_rpm/tests/performance/test_sync.py::SyncTestCase::test_centos8_appstream_on_demand
157.77s call     pulp_rpm/tests/performance/test_sync.py::SyncTestCase::test_centos8_kickstart_baseos_on_demand
128.60s call     pulp_rpm/tests/performance/test_sync.py::SyncTestCase::test_centos8_kickstart_appstream_on_demand
0.38s setup    pulp_rpm/tests/performance/test_sync.py::SyncTestCase::test_centos7_on_demand
0.00s teardown pulp_rpm/tests/performance/test_sync.py::SyncTestCase::test_centos8_kickstart_baseos_on_demand
0.00s setup    pulp_rpm/tests/performance/test_sync.py::SyncTestCase::test_centos8_appstream_on_demand
0.00s setup    pulp_rpm/tests/performance/test_sync.py::SyncTestCase::test_centos8_kickstart_appstream_on_demand
0.00s setup    pulp_rpm/tests/performance/test_sync.py::SyncTestCase::test_centos8_baseos_on_demand
0.00s setup    pulp_rpm/tests/performance/test_sync.py::SyncTestCase::test_centos8_kickstart_baseos_on_demand
0.00s teardown pulp_rpm/tests/performance/test_sync.py::SyncTestCase::test_centos8_baseos_on_demand
0.00s teardown pulp_rpm/tests/performance/test_sync.py::SyncTestCase::test_centos7_on_demand
0.00s teardown pulp_rpm/tests/performance/test_sync.py::SyncTestCase::test_centos8_appstream_on_demand
0.00s teardown pulp_rpm/tests/performance/test_sync.py::SyncTestCase::test_centos8_kickstart_appstream_on_demand
=============================================================================== 5 passed in 975.89s (0:16:15) ================================================================================

```